requirements.txt
aniso8601==1.1.0
beautifulsoup4==4.4.1
coverage==3.7.1
cssselect==0.9.1
HTMLParser==0.0.2
itsdangerous==0.24
Jinja2==2.8
lxml==3.5.0
MarkupSafe==0.23
mechanize==0.2.5
nose==1.3.7
python-coveralls==2.5.0
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
readability-lxml==0.6.1
redis==2.10.5
requests==2.8.1
sh==1.11
six==1.10.0
Werkzeug==0.11.2
wheel==0.24.0
